POST requests: remove extraneous copying
`post_max_size` limits the size of the traditional upload process used
by older browsers. We'd like to have that set as high as possible,
BUT...

If php receives a POST request that is less than `post_max_size` and
it comes from a modern browser (the content type is anything other
than `multipart/form-data`), php would previously run inneficient code
that copied the request body several times for backward compatability
reasons.  This needed roughly `3 * $filesize` bytes of on-heap memory
for a given upload.  This effectively is a tradeoff, set
`post_max_size` high and modern browsers start making php run out of
memory, set it low and older browsers can't upload big files.

A new commit in php-5.6 removes this old code along with a bunch of
other stuff. See:
https://github.com/php/php-src/commit/2438490addfbfba51e12246a74588b2382caa08a
We can't jut apply this patch whole-sale cause we don't know what else
it depends on. Eventually we'll use 5.6 and this change won't be
necessary.

This simply removes the old code that copies the body several times
and only restores `raw_post_data` (for reading from stdin) when one of
php's content type parsers reads the whole body.